Movies project
a small project I have built over the last 2 days, it asks the user for
the title, length, director, year of release, and genre of a movie and
adds it to a database.  I'm looking to deal with some of the strange tab
behavior and am hopefully going to add in the ability to get info from
IMDB soon.